Lab5/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=10.0.0
requests>=2.31.0
git+https://github.com/openai/CLIP.git